Move content to model * consolidate into model_fn * Fix cancellation error
